removing alt-x from skipLink into its own custom hook and using in patient search
