+++ apps/modernization-ui/src/apps/search/advancedSearch/AdvancedSearch.tsx
@@ -76,12 +76,7 @@ export const AdvancedSearch = () => {
     const [resultStartCount, setResultStartCount] = useState<number>(0);
     const [resultEndCount, setResultEndCount] = useState<number>(0);
     const [resultTotal, setResultTotal] = useState<number>(0);
-    const { skipTo, previousTarget } = useSkipLink();
-
-    useEffect(() => {
-        skipTo('lastName');
-        previousTarget('lastName');
-    }, []);
+    const { skipTo } = useSkipLink();
 
     const [showAddNewDropDown, setShowAddNewDropDown] = useState<boolean>(false);
     const [
